+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBQueriedRecord.java
@@ -29,6 +29,7 @@ import com.google.protobuf.Message;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.util.Map;
 
 /**
  * A record returned by a query and therefore possibly associated with a particular entry in some index.
@@ -61,6 +62,13 @@ public abstract class FDBQueriedRecord<M extends Message> implements FDBRecord<M
     @Nullable
     public abstract IndexEntry getIndexEntry();
 
+    /**
+     * The constituents of this record if it is a synthetic record type.
+     * @return a map of constituents from name to {@link FDBStoredRecord}
+     */
+    @Nullable
+    public abstract Map<String, FDBStoredRecord<? extends Message>> getConstituents();
+
     public static <M extends Message> FDBQueriedRecord<M> indexed(@Nonnull FDBIndexedRecord<M> indexed) {
         return new Indexed<>(indexed);
     }
@@ -71,6 +79,10 @@ public abstract class FDBQueriedRecord<M extends Message> implements FDBRecord<M
 
     public static <M extends Message> FDBQueriedRecord<M> covered(@Nonnull Index index, @Nonnull IndexEntry indexEntry, @Nonnull Tuple primaryKey, @Nonnull RecordType recordType, @Nonnull M protoRecord) {
         return new Covered<>(index, indexEntry, primaryKey, recordType, protoRecord);
+    }
+
+    public static <M extends Message> FDBQueriedRecord<M> synthetic(@Nonnull Index index, @Nonnull IndexEntry indexEntry, @Nonnull FDBSyntheticRecord syntheticRecord) {
+        return new Synthetic<>(index, indexEntry, syntheticRecord.getPrimaryKey(), syntheticRecord.getRecordType(), syntheticRecord.getRecord(), syntheticRecord.getConstituents());
     }
 
     @SuppressWarnings("PMD.AvoidFieldNameMatchingTypeName")
@@ -126,6 +138,12 @@ public abstract class FDBQueriedRecord<M extends Message> implements FDBRecord<M
         @Override
         public IndexEntry getIndexEntry() {
             return indexed.getIndexEntry();
+        }
+
+        @Nullable
+        @Override
+        public Map<String, FDBStoredRecord<? extends Message>> getConstituents() {
+            return null;
         }
     }
 
@@ -183,6 +201,12 @@ public abstract class FDBQueriedRecord<M extends Message> implements FDBRecord<M
         public IndexEntry getIndexEntry() {
             return null;
         }
+
+        @Nullable
+        @Override
+        public Map<String, FDBStoredRecord<? extends Message>> getConstituents() {
+            return null;
+        }
     }
 
     static class Covered<M extends Message> extends FDBQueriedRecord<M> {
@@ -205,7 +229,7 @@ public abstract class FDBQueriedRecord<M extends Message> implements FDBRecord<M
             this.protoRecord = protoRecord;
         }
 
-        @Nullable
+        @Nonnull
         @Override
         public Index getIndex() {
             return index;
@@ -251,6 +275,99 @@ public abstract class FDBQueriedRecord<M extends Message> implements FDBRecord<M
         public FDBStoredRecord<M> getStoredRecord() {
             return null;
         }
+
+        @Nullable
+        @Override
+        public Map<String, FDBStoredRecord<? extends Message>> getConstituents() {
+            return null;
+        }
     }
 
+    static class Synthetic<M extends Message> extends FDBQueriedRecord<M> {
+        @Nonnull
+        private final Index index;
+        @Nonnull
+        private final IndexEntry indexEntry;
+        @Nonnull
+        private final Tuple primaryKey;
+        @Nonnull
+        private final RecordType recordType;
+        @Nonnull
+        private final Message protoRecord;
+        @Nonnull
+        private final Map<String, FDBStoredRecord<? extends Message>> constituents;
+
+        public Synthetic(@Nonnull final Index index,
+                         @Nonnull final IndexEntry indexEntry,
+                         @Nonnull final Tuple primaryKey,
+                         @Nonnull final RecordType recordType,
+                         @Nonnull final Message protoRecord,
+                         @Nonnull final Map<String, FDBStoredRecord<? extends Message>> constituents) {
+            this.index = index;
+            this.indexEntry = indexEntry;
+            this.primaryKey = primaryKey;
+            this.recordType = recordType;
+            this.protoRecord = protoRecord;
+            this.constituents = constituents;
+        }
+
+        @Nonnull
+        @Override
+        public Index getIndex() {
+            return index;
+        }
+
+        @Nonnull
+        @Override
+        public IndexEntry getIndexEntry() {
+            return indexEntry;
+        }
+
+        @Nonnull
+        @Override
+        public Tuple getPrimaryKey() {
+            return primaryKey;
+        }
+
+        @Nonnull
+        @Override
+        public RecordType getRecordType() {
+            return recordType;
+        }
+
+        /**
+         * Returns the associated protobuf message as object of type {@code M}. This will do an unchecked cast.
+         * It is the responsibility of the caller to not call this method using a typed store.
+         * @return the message associated with this record
+         */
+        @Nonnull
+        @Override
+        @SuppressWarnings("unchecked")
+        public M getRecord() {
+            return (M)protoRecord;
+        }
+
+        @Override
+        public boolean hasVersion() {
+            return false;
+        }
+
+        @Nullable
+        @Override
+        public FDBRecordVersion getVersion() {
+            return null;
+        }
+
+        @Nullable
+        @Override
+        public FDBStoredRecord<M> getStoredRecord() {
+            return null;
+        }
+
+        @Nonnull
+        @Override
+        public Map<String, FDBStoredRecord<? extends Message>> getConstituents() {
+            return constituents;
+        }
+    }
 }